spielzeitenplaner/src/main/java/de/bathesis/spielzeitenplaner/web/controller/SettingsController.java
package de.bathesis.spielzeitenplaner.web.controller;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;

import de.bathesis.spielzeitenplaner.domain.Formation;
import de.bathesis.spielzeitenplaner.mapper.FormationMapper;
import de.bathesis.spielzeitenplaner.services.SettingsService;
import de.bathesis.spielzeitenplaner.web.forms.FormationForm;


@Controller
@RequestMapping("/settings")
public class SettingsController {

    private final SettingsService settingsService;

    public SettingsController(SettingsService settingsService) {
        this.settingsService = settingsService;
    }


    @GetMapping
    public String settings() {
        return "settings";
    }

    @PostMapping("/saveFormation")
    public String saveFormation(FormationForm formationForm) {
        Formation formation = FormationMapper.toDomainFormation(formationForm);
        settingsService.save(formation);
        return "redirect:/settings";
    }

}
